move methods to other resolvers
