+++ service/src/main/java/org/cbioportal/genome_nexus/service/internal/BaseVariantAnnotationServiceImpl.java
@@ -34,7 +34,8 @@ package org.cbioportal.genome_nexus.service.internal;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.cbioportal.genome_nexus.component.annotation.IndexResolver;
+import org.cbioportal.genome_nexus.component.annotation.HugoGeneSymbolResolver;
+import org.cbioportal.genome_nexus.component.annotation.IndexBuilder;
 import org.cbioportal.genome_nexus.component.annotation.ProteinChangeResolver;
 import org.cbioportal.genome_nexus.model.*;
 import org.cbioportal.genome_nexus.persistence.IndexRepository;
@@ -73,6 +74,7 @@ public abstract class BaseVariantAnnotationServiceImpl implements VariantAnnotat
     private final ClinvarVariantAnnotationService clinvarVariantAnnotationService;
     private final IndexRepository indexRepository;
     private final ProteinChangeResolver proteinChangeResolver;
+    private final HugoGeneSymbolResolver hugoGeneSymbolResolver;
 
     public BaseVariantAnnotationServiceImpl(
         BaseCachedExternalResourceFetcher<VariantAnnotation, VariantAnnotationRepository> resourceFetcher,
@@ -87,7 +89,8 @@ public abstract class BaseVariantAnnotationServiceImpl implements VariantAnnotat
         OncokbService oncokbService,
         ClinvarVariantAnnotationService clinvarVariantAnnotationService,
         IndexRepository indexRepository,
-        ProteinChangeResolver proteinChangeResolver
+        ProteinChangeResolver proteinChangeResolver,
+        HugoGeneSymbolResolver hugoGeneSymbolResolver
     ) {
         this.resourceFetcher = resourceFetcher;
         this.ensemblService = ensemblService;
@@ -102,6 +105,7 @@ public abstract class BaseVariantAnnotationServiceImpl implements VariantAnnotat
         this.clinvarVariantAnnotationService = clinvarVariantAnnotationService;
         this.indexRepository = indexRepository;
         this.proteinChangeResolver = proteinChangeResolver;
+        this.hugoGeneSymbolResolver = hugoGeneSymbolResolver;
     }
 
     // Needs to be overridden to support normalizing variants
@@ -110,9 +114,9 @@ public abstract class BaseVariantAnnotationServiceImpl implements VariantAnnotat
         return id;
     }
 
-    private Index extractIndex(VariantAnnotation variantAnnotation) {
-        IndexResolver resolver = new IndexResolver(this.proteinChangeResolver);
-        return resolver.resolve(variantAnnotation);
+    private Index buildIndex(VariantAnnotation variantAnnotation) {
+        IndexBuilder builder = new IndexBuilder(this.proteinChangeResolver, this.hugoGeneSymbolResolver);
+        return builder.buildIndex(variantAnnotation);
     }
 
     @Override
@@ -164,9 +168,7 @@ public abstract class BaseVariantAnnotationServiceImpl implements VariantAnnotat
             
             // add new annotation to index db
             variantAnnotation.ifPresent(annotation -> {
-                Gson gson = new Gson();
-                DBObject dbObject = BasicDBObject.parse(gson.toJson(this.extractIndex(annotation)));
-                this.indexRepository.saveDBObject("index", normalizedVariant, dbObject);
+                this.saveToIndexDb(normalizedVariant, annotation);
             });
     
             // include original variant value too
@@ -207,10 +209,7 @@ public abstract class BaseVariantAnnotationServiceImpl implements VariantAnnotat
             variantAnnotations = this.resourceFetcher.fetchAndCache(new ArrayList(normVarToOrigVarQueryMap.keySet()));
             for (VariantAnnotation variantAnnotation : variantAnnotations) {
                 // add new annotation to index db
-                Gson gson = new Gson();
-                DBObject dbObject = BasicDBObject.parse(gson.toJson(this.extractIndex(variantAnnotation)));
-                this.indexRepository.saveDBObject("index", normVarToOrigVarQueryMap.get(variantAnnotation.getVariant()), dbObject);
-                
+                this.saveToIndexDb(normVarToOrigVarQueryMap.get(variantAnnotation.getVariant()), variantAnnotation);
                 variantAnnotation.setOriginalVariantQuery(normVarToOrigVarQueryMap.get(variantAnnotation.getVariant()));
             }
         } catch (HttpClientErrorException e) {
@@ -269,6 +268,12 @@ public abstract class BaseVariantAnnotationServiceImpl implements VariantAnnotat
         }
 
         return variantAnnotations;
+    }
+
+    private void saveToIndexDb(String normalizedVariant, VariantAnnotation annotation) {
+        Gson gson = new Gson();
+        DBObject dbObject = BasicDBObject.parse(gson.toJson(this.buildIndex(annotation)));
+        this.indexRepository.saveDBObject("index", normalizedVariant, dbObject);
     }
