+++ service/src/main/java/org/cbioportal/genome_nexus/service/internal/HgvsVariantAnnotationService.java
@@ -35,6 +35,7 @@ package org.cbioportal.genome_nexus.service.internal;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.cbioportal.genome_nexus.service.*;
+import org.cbioportal.genome_nexus.component.annotation.HugoGeneSymbolResolver;
 import org.cbioportal.genome_nexus.component.annotation.NotationConverter;
 import org.cbioportal.genome_nexus.component.annotation.ProteinChangeResolver;
 import org.cbioportal.genome_nexus.persistence.IndexRepository;
@@ -66,7 +67,8 @@ public class HgvsVariantAnnotationService extends BaseVariantAnnotationServiceIm
         @Lazy OncokbService oncokbService,
         @Lazy ClinvarVariantAnnotationService clinvarVariantAnnotationService,
         IndexRepository indexRepository,
-        ProteinChangeResolver proteinChangeResolver
+        ProteinChangeResolver proteinChangeResolver,
+        HugoGeneSymbolResolver hugoGeneSymbolResolver
     ) {
         super(
             cachedVariantAnnotationFetcher,
@@ -81,7 +83,8 @@ public class HgvsVariantAnnotationService extends BaseVariantAnnotationServiceIm
             oncokbService,
             clinvarVariantAnnotationService,
             indexRepository,
-            proteinChangeResolver
+            proteinChangeResolver,
+            hugoGeneSymbolResolver
         );
 
         this.notationConverter = notationConverter;
